Refactor: Enum의 책임을 줄이기 위해 Create구조 변경
