--- src/main/java/org/prgrms/kdtspringhomework/command/create/CreateVoucherStrategy.java
@@ -1,8 +1,0 @@
-package org.prgrms.kdtspringhomework.command.create;
-
-import org.prgrms.kdtspringhomework.voucher.service.VoucherService;
-
-@FunctionalInterface
-public interface CreateVoucherStrategy {
-    boolean create(VoucherService voucherService, long amount);
-}
