+++ src/main/java/org/prgrms/kdtspringhomework/voucher/service/VoucherService.java
@@ -1,5 +1,6 @@
 package org.prgrms.kdtspringhomework.voucher.service;
 
+import org.prgrms.kdtspringhomework.command.CreateVoucherStatus;
 import org.prgrms.kdtspringhomework.voucher.domain.Voucher;
 import org.prgrms.kdtspringhomework.voucher.repository.VoucherRepository;
 
@@ -19,7 +20,8 @@ public class VoucherService {
                 .orElseThrow(() -> new RuntimeException(String.format("Can not find a voucher for %s", voucherId)));
     }
 
-    public void addVoucher(Voucher voucher) {
+    public void addVoucher(String voucherType, long amount) {
+        Voucher voucher = CreateVoucherStatus.getVoucher(voucherType, UUID.randomUUID(), amount);
         voucherRepository.add(voucher);
     }
